moved remaining eprog2/3 methods to eprog simplified Token
