+++ pol-core/bscript/token.h
@@ -49,29 +49,20 @@ public:
     const unsigned char* dataptr;
   };
   bool deprecated;
-  bool ownsStr;
   unsigned char module;
 
   static unsigned int instances();
   static void show_instances();
 
 protected:
-  const char* token;
+  std::string token;
 
 public:
-  const char* tokval() const { return token; }
+  const char* tokval() const { return token.c_str(); }
   Token();
-  Token( const Token& tok );
-  Token& operator=( const Token& tok );
 
-  Token( ModuleID module, BTokenId id, BTokenType type );
-  Token( BTokenId id, BTokenType type );
   void nulStr();
-  void setStr( const char* s );
-  void copyStr( const char* s );
-  void copyStr( const char* s, int len );
-
-  ~Token();
+  void setStr( std::string str );
 
   void printOn( std::ostream& outputStream ) const;
 };
