+++ pol-core/bscript/tokens.h
@@ -313,13 +313,5 @@ inline auto format_as( BTokenId t )
   return fmt::underlying( t );
 }
 }  // namespace Bscript
-
-namespace Clib
-{
-template <>
-std::string tostring<Bscript::BTokenType>( const Bscript::BTokenType& v );
-template <>
-std::string tostring<Bscript::BTokenId>( const Bscript::BTokenId& v );
-}  // namespace Clib
 }  // namespace Pol
 #endif
